requirements.txt
colorlog==6.10.1
coppeliasim_zmqremoteapi_client==2.0.4
matplotlib==3.7.2
numpy==1.24.3
open3d==0.19.0
opencv_python==4.12.0.88
Pillow==10.0.1
Pillow==12.1.1
protobuf==6.33.5
python-dotenv==1.2.1
setuptools==80.9.0
setuptools==68.0.0
unified_planning==1.3.0
